LayoutTests/http/tests/security/cross-frame-access-first-time-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-get-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-first-time.html. Domains, protocols and ports must match.

This test checks cross-frame access security for first-time access to the document (rdar://problem/5251309).


canGet(frames[0], 'document'): false
canGet(frames[0].document, 'cookie'): false
canGet(frames[0].document, 'body'): false
canGet(frames[0].document, 'location'): false
